+++ influxdb3-plugin-sdk/src/mutate_index.rs
@@ -125,6 +125,7 @@ mod tests {
             dependencies: Dependencies {
                 database_version: ">=3.0.0".parse().unwrap(),
                 python: vec![],
+                plugins: vec![],
             },
             hash: ArtifactHash::try_new(
                 "sha256:0000000000000000000000000000000000000000000000000000000000000000",